backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.1
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
django-cors-headers==3.7.0
    # via -r requirements.in
django==3.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
idna==2.10
    # via requests
pytz==2020.5
    # via django
requests==2.25.1
    # via
    #   -r requirements.in
    #   stripe
sqlparse==0.4.1
    # via django
stripe==2.55.1
    # via -r requirements.in
urllib3==1.26.2
    # via requests
